add images to every page with plants
