+++ components/AddPlantForm/AddPlantForm.js
@@ -14,38 +14,7 @@ import {
   StyledErrorH3,
   StyledLoading,
 } from "../RoomList/RoomsList.Styled";
-
-export const PlantType = Object.freeze({
-  INVALID: "INVALID",
-  MONEY_PLANT: "MONEY_PLANT",
-  PHILODENDRON: "PHILODENDRON",
-  CHINESE_EVERGREEN: "CHINESE_EVERGREEN",
-  SPIDER_PLANT: "SPIDER_PLANT",
-  STRING_OF_PERLS: "STRING_OF_PERLS",
-  STAGHORN_FERN: "STAGHORN_FERN",
-  ORCHID: "ORCHID",
-  PEACE_LILY: "PEACE_LILY",
-  MONSTERA_DELICIOSA: "MONSTERA_DELICIOSA",
-  SNAKE_PLANT: "SNAKE_PLANT",
-  LILY_OF_THE_VALLEY: "LILY_OF_THE_VALLEY",
-  JADE_PLANT: "JADE_PLANT",
-  STING_OF_HEARTS: "STING_OF_HEARTS",
-  RUBBER_PLANT: "RUBBER_PLANT",
-  ASPARAGUS_FERN: "ASPARAGUS_FERN",
-  CALATHEA_ORNATA: "CALATHEA_ORNATA",
-  YUCCA_PLANT: "YUCCA_PLANT",
-  ENGLISH_IVY: "ENGLISH_IVY",
-  DRACAENA_MARGINATA: "DRACAENA_MARGINATA",
-  PEPEROMIA: "PEPEROMIA",
-  ZZ_PLANT: "ZZ_PLANT",
-  ARROWHEAD_VINE: "ARROWHEAD_VINE",
-  HOYA: "HOYA",
-  RUBBER_TREE: "RUBBER_TREE",
-  CAST_IRON_PLANT: "CAST-IRON_PLANT",
-  CORN_PLANT: "CORN_PLANT",
-  PONYTAIL_PALM: "PONYTAIL_PALM",
-  CALATHEA: "CALATHEA",
-});
+import { PlantType } from "@/utils/PlantType";
 
 export default function AddPlantForm() {
   const { mutate } = useSWRConfig();
@@ -130,7 +99,7 @@ export default function AddPlantForm() {
           <option value={PlantType.INVALID}>No plant type</option>
           <option value={PlantType.MONEY_PLANT}>Money Plant</option>
           <option value={PlantType.PHILODENDRON}>Philodendron</option>
-          <option value={PlantType.CHINESE_EVERGREEN}>Chinese evergreen</option>
+          <option value={PlantType.CHINESE_EVERGREEN}>Chinese Evergreen</option>
           <option value={PlantType.SPIDER_PLANT}>Spider Plant</option>
           <option value={PlantType.STRING_OF_PERLS}>String of perls</option>
           <option value={PlantType.STAGHORN_FERN}>Staghorn fern</option>
