+++ components/AllPlantsList/AllPlantsList.js
@@ -10,6 +10,7 @@ import {
   StyledErrorH3,
   StyledLoading,
 } from "../RoomList/RoomsList.Styled";
+import { StyledImage } from "@/pages";
 
 export default function AllPlantsList({ room }) {
   const { data: plants, error, isLoading } = useSWR(`/api/plants`, fetcher);
@@ -32,7 +33,14 @@ export default function AllPlantsList({ room }) {
         {plants.map((plant) => (
           <StyledCard key={plant._id}>
             <StyledListItem key={plant._id}>
-              🪴 {plant.title} <br /> Room:
+              <StyledImage
+                key={plant._id}
+                src={`/plants/${plant.type}.jpeg`}
+                width={90}
+                height={90}
+                alt={plant.title}
+              ></StyledImage>{" "}
+              {plant.title} <br /> Room:
               {plant.rooms.length > 0
                 ? plant.rooms[0].title
                 : "No room assigned"}
